ejeweeka_backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pgvector>=0.2.4
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-genai>=0.6.0
python-multipart>=0.0.6
httpx>=0.25.0
aiogram>=3.4.0
asyncpg>=0.29.0
slowapi>=0.1.9
sentry-sdk[fastapi]>=1.40.0
